engines/cortex/requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
cortex4py==1.0.0
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
json2html==1.2.1
MarkupSafe==1.0
python-magic==0.4.15
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
